% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{batch_frequency_to_dataframe}
\alias{batch_frequency_to_dataframe}
\title{batch_frequency_to_dataframe}
\usage{
batch_frequency_to_dataframe(relative = TRUE)
}
\arguments{
\item{relative}{relative = true ; false = absolute}
}
\value{
dataframe that contains frequency information
}
\description{
batch_frequency_to_dataframe
}
