% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{batch_size_stats_as_data_frame}
\alias{batch_size_stats_as_data_frame}
\title{batch_size_stats_as_data_frame}
\usage{
batch_size_stats_as_data_frame()
}
\value{
data frame for shiny app that shows for every batch type the corresponiding activity and its size metrics(mean , median)
}
\description{
batch_size_stats_as_data_frame
}
