% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{batch_waiting_time_to_data_frame}
\alias{batch_waiting_time_to_data_frame}
\title{batch_waiting_time_to_data_frame}
\usage{
batch_waiting_time_to_data_frame()
}
\value{

}
\description{
batch_waiting_time_to_data_frame
}
