% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchanalytics-package.R
\docType{package}
\name{batchanalytics-package}
\alias{batchanalytics}
\alias{batchanalytics-package}
\title{batchanalytics: Provides Functions That Represent Metrics For Analysing Batching Behavior.}
\description{
Provides Functions For Analysing Batching Behavior. The Package is build as an extension for the bama algorihm [https://github.com/nielsmartin/bama]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NiklasCarlos1994/batchanalytics}
  \item Report bugs at \url{https://github.com/NiklasCarlos1994/batchanalytics/issues}
}

}
\keyword{internal}
