% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{compare_processing_time_of_activites}
\alias{compare_processing_time_of_activites}
\title{compare processing time for each batching type and see activity duration}
\usage{
compare_processing_time_of_activites()
}
\value{

}
\description{
compare processing time for each batching type and see activity duration
}
