% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{create_recommendations}
\alias{create_recommendations}
\title{create_recommendations}
\usage{
create_recommendations(metricStats)
}
\arguments{
\item{metricStats}{}
}
\value{
recommendation for specific metric how to proceed with process as string
}
\description{
create_recommendations
}
