% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_act_for_specific_batch_type}
\alias{get_act_for_specific_batch_type}
\title{get_act_for_specific_batch_type}
\usage{
get_act_for_specific_batch_type(my_batch_type)
}
\arguments{
\item{my_batch_type}{}
}
\value{
gets for a specific batch types all the activites
}
\description{
get_act_for_specific_batch_type
}
