% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_batch_types}
\alias{get_batch_types}
\title{get_batch_types}
\usage{
get_batch_types(res_log)
}
\arguments{
\item{res_log}{}
}
\value{
returns batch types that are found in the result log
}
\description{
get_batch_types
}
