% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_batching_activities_for_each_type}
\alias{get_batching_activities_for_each_type}
\title{show_batching_activities_for_each_type}
\usage{
get_batching_activities_for_each_type()
}
\value{
activities that use batching in a list first element of list is sim, 2. seq, 3. conc
}
\description{
which activities use batching
}
