% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_batching_df_logs}
\alias{get_batching_df_logs}
\title{get batching dataframes}
\usage{
get_batching_df_logs(res_log)
}
\arguments{
\item{res_Log}{is the log which results form}
}
\value{
creates global vars that contain different batching infos and can be used for further analysis
}
\description{
get batching dataframes
}
