% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_idle_time}
\alias{get_idle_time}
\title{get_idle_time
helper FUN for getting idle time(solve the elog cant be empty problem with replacing null)}
\usage{
get_idle_time(elog)
}
\arguments{
\item{elog}{}
}
\value{
idle time
}
\description{
get_idle_time
helper FUN for getting idle time(solve the elog cant be empty problem with replacing null)
}
