% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{get_metric_stats}
\alias{get_metric_stats}
\title{get_metric_stats}
\usage{
get_metric_stats(metric_fun)
}
\arguments{
\item{metric_fun}{metric function like compare_processing_time(bplot = FALSE) as example}
}
\value{
returns the median of all batch types and safes in a c(parallel,seq,conc, noBatch)
}
\description{
get_metric_stats
}
