% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello world test function}
\usage{
hello(x)
}
\arguments{
\item{x}{word that gets concatenate with hello...}
}
\value{
output form print -< two pasted strings
}
\description{
My hello world test function
}
