% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{metric_activity_duration}
\alias{metric_activity_duration}
\title{Title}
\usage{
metric_activity_duration(activity_log_with_batches, act, res, type_of_batch)
}
\arguments{
\item{type_of_batch}{}
}
\value{

}
\description{
Title
}
