% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{metric_batch_size}
\alias{metric_batch_size}
\title{Title}
\usage{
metric_batch_size(
  activity_log_with_batches,
  act,
  res,
  type_of_batch,
  exclude_singletons = TRUE
)
}
\arguments{
\item{exclude_singletons}{}
}
\value{

}
\description{
Title
}
