% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{metric_frequency}
\alias{metric_frequency}
\title{Metric - Batch processing frequency}
\usage{
metric_frequency(
  activity_log_with_batches,
  act,
  res,
  type_of_batch,
  relative = FALSE
)
}
\arguments{
\item{relative}{}
}
\value{

}
\description{
Metric - Batch processing frequency
}
