% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{metric_n_concurrent_cases}
\alias{metric_n_concurrent_cases}
\title{Title}
\usage{
metric_n_concurrent_cases(activity_log_with_batches, act, res)
}
\arguments{
\item{res}{}
}
\value{

}
\description{
Title
}
