% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{metric_waiting_time}
\alias{metric_waiting_time}
\title{Title}
\usage{
metric_waiting_time(activity_log_with_batches, act, res, type_of_batch)
}
\arguments{
\item{type_of_batch}{}
}
\value{

}
\description{
Title
}
