% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{my_detect_batching}
\alias{my_detect_batching}
\title{my_detect_batching
function to facilitate the batching identification process}
\usage{
my_detect_batching(task_log)
}
\arguments{
\item{task_log}{}
}
\value{
result log with batching behaviour
}
\description{
my_detect_batching
function to facilitate the batching identification process
}
