% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{shiny_metric_batch_frequency}
\alias{shiny_metric_batch_frequency}
\title{shiny_metric_batch_frequency}
\usage{
shiny_metric_batch_frequency(
  activity_log_with_batches,
  act,
  type_of_batch,
  relative = TRUE
)
}
\arguments{
\item{relative}{}
}
\value{

}
\description{
shiny_metric_batch_frequency
}
