% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{shiny_metric_batch_size}
\alias{shiny_metric_batch_size}
\title{batch size fun optimized for use in shiny app}
\usage{
shiny_metric_batch_size(
  activity_log_with_batches,
  act,
  type_of_batch,
  exclude_singletons = TRUE
)
}
\value{

}
\description{
batch size fun optimized for use in shiny app
}
