% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{shiny_metric_waiting_time}
\alias{shiny_metric_waiting_time}
\title{shiny_metric_waiting_time}
\usage{
shiny_metric_waiting_time(activity_log_with_batches, act, type_of_batch)
}
\arguments{
\item{type_of_batch}{}
}
\value{

}
\description{
shiny_metric_waiting_time
}
