% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{show_batching_activities}
\alias{show_batching_activities}
\title{show_batching_activities}
\usage{
show_batching_activities()
}
\value{
activities that use batching
}
\description{
which activities use batching
}
