% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{show_batching_in_process_map}
\alias{show_batching_in_process_map}
\title{show batching in process map}
\usage{
show_batching_in_process_map()
}
\value{

}
\description{
show batching in process map
}
