% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{show_dataFrame_with_batching_activities}
\alias{show_dataFrame_with_batching_activities}
\title{show_dataFrame_with_batching_activities}
\usage{
show_dataFrame_with_batching_activities(liste)
}
\value{
activities that use batching
}
\description{
show dataframe with  activities use batching
}
\details{
get_batching_activities_for_each_type methode is usually the liste parameter
}
