% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{show_most_eff_act}
\alias{show_most_eff_act}
\title{show_most_eff_act}
\usage{
show_most_eff_act()
}
\value{
df with most efficient batching behaviour for each activity
}
\description{
show_most_eff_act
}
