% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBatchAnalytics.R
\name{transform_df_to_event_log}
\alias{transform_df_to_event_log}
\title{transform data frame to event log}
\usage{
transform_df_to_event_log()
}
\value{
creates global vars elogs for each batching type for further analysis
}
\description{
transform data frame to event log
}
